Separated out IAMservice code from WebServer file
