+++ iamserver/iam_settings.hpp
@@ -32,9 +32,9 @@ public:
 	uint32_t getUserIdxOffset() const {
 		return useridx_offset;
 	}
-	std::string getLoginPageContent() const {
-		std::string sLoginPage(login_content);
-		return sLoginPage;
+	std::string getAuthPageContent() const {
+		std::string sAuthPage(auth_content);
+		return sAuthPage;
 	}
 	/**
 	 * Set relevant values
@@ -90,7 +90,7 @@ public:
 private:
 	bool is_enabled_{ false };
 	uint32_t useridx_offset = OAUTH2_USERIDX_OFFSET;
-	const char *login_content = 
+	const char *auth_content =
 	#include "views/iam_auth.html"
 	;
 };
